+++ cvat-cli/src/cvat_cli/_internal/commands_profile.py
@@ -6,35 +6,13 @@ from __future__ import annotations
 
 import argparse
 import getpass
-import json
 from pathlib import Path
 
 from cvat_sdk.core.auth import DEFAULT_SERVER, AuthStore, ProfileEntry
 
 from .command_base import CommandGroup
 from .common import CriticalError
-from .utils import _fetch_name_from_server, _normalize_server, _now_iso
-
-
-def _read_token_file(path: Path) -> tuple[str, str | None, str | None]:
-    """Read a PAT from *path*, returning (token, envelope_server, envelope_name).
-
-    Accepts two file shapes:
-
-    * plain text - the whole file (whitespace-trimmed) is the token.
-    * JSON envelope emitted by the web UI - ``{"version": 1, "server": ...,
-      "name": ..., "token": ...}``; the envelope's ``server`` and ``name``
-      participate in resolution but are overridden by explicit CLI flags.
-    """
-    text = path.read_text(encoding="utf-8")
-    try:
-        doc = json.loads(text)
-    except json.JSONDecodeError:
-        return text.strip(), None, None
-
-    if not isinstance(doc, dict) or "token" not in doc:
-        return text.strip(), None, None
-    return doc["token"], doc.get("server"), doc.get("name")
+from .utils import _fetch_name_from_server, _normalize_server, _now_iso, _read_token_file
 
 COMMANDS = CommandGroup(description="Manage saved CVAT authentication profiles.")
